migrate LoraFunction from tudb:main
